pyproject.toml
[tool.poetry]
name = "seo"
version = "0.1.0"
description = ""
authors = ["Ababacar BA <ababacar.ba@openvalue.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = "2.2.2"
langchain = "0.1.1"
langchain-community = "0.0.14"
langchain-core = "0.1.14"
openai = "1.34.0"
unstructured = "^0.15.0"
streamlit-pdf-viewer = "^0.0.14"
azure-identity = "1.16.1"
azure-keyvault = "4.2.0"
chromadb = "0.4.22"
streamlit = "^1.38.0"
rank-bm25 = "0.2.2"
streamlit-feedback = "0.1.3"
openpyxl = "3.1.2"
gpt4all = "2.6.0"
sentence-transformers = "2.7.0"
pyyaml = "6.0.1"
python-dotenv = "1.0.1"
pandas = "2.2.2"
pypdf = "^4.3.1"
opencv-python = "^4.10.0.84"
pillow-heif = "^0.17.0"
pdf2image = "^1.17.0"
azure-storage-blob = "^12.21.0"
azure-blobstorage-utils = {git = "https://github.com/OpenValue/azure_blobstorage_utils.git"}
googlesearch-python = "^1.2.5"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
jupyter = "^1.1.1"
jupyter-black = "^0.4.0"
ipykernel = "^6.29.5"
notebook = "^7.2.2"
jupyter-contrib-nbextensions = "0.5.1"
jupyter-nbextensions-configurator = "0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''